Song Tones implementation and revisions
Made song tones non-blocking, faster, and more efficient. They're been turned from a .ino to a .cpp and a .h to be referenced by the main .ino file. Implemented it into the OS better by having the saved song play when the alarm goes off and having a settings menu for choosing the song you want.